tweak: change task kill to use script PID instead of executable
this makes it so that if the task is run again, it doesn't kill the running pid of the command, but rather the pid of the script itself, which vscode does not see as an error and doesn't show a error popup

this means that it could kill itself, so a new check was added to the bash condition

+++ api/.vscode/scripts/debug-api.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 
-API_DEV_PID=$(pgrep -f 'pnpm start:debug' -d ' ')
+API_DEV_PID=$(pgrep -f 'debug-api.sh' -d ' ')
 
 # If there is a api dev server already running, kill it.
-if [ ! -z $API_DEV_PID ]
+if [ ! -z $API_DEV_PID ] && [ $API_DEV_PID != $$ ]
 then
 	kill $API_DEV_PID
 fi
