+++ client/.vscode/scripts/debug-client.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 
-CLIENT_DEV_PID=$(pgrep -f 'vite\.js dev' -d ' ')
+CLIENT_DEV_PID=$(pgrep -f 'debug-client.sh' -d ' ')
 
 # If there is a vite dev server already running, kill it.
-if [ ! -z $CLIENT_DEV_PID ]
+if [ ! -z $CLIENT_DEV_PID ] && [ $CLIENT_DEV_PID != $$ ]
 then
 	kill $CLIENT_DEV_PID
 fi
